+++ module/helios-generique/template/HeliosSignatureLocale.php
@@ -21,16 +21,24 @@ $libersignConnecteur->displayLibersignJS();
             siginfos.push({
                 hash: "<?php echo ($signatureInfo['isbordereau'] == true) ? $signatureInfo['bordereau_hash'] : $signatureInfo['flux_hash'] ?>",
                 pesid: "<?php echo ($signatureInfo['isbordereau'] == true) ? $signatureInfo['bordereau_id'] : $signatureInfo['flux_id'] ?>",
-                pespolicyid: "urn:oid:1.2.250.1.131.1.5.18.21.1.4",
+                //pespolicyid: "urn:oid:1.2.250.1.131.1.5.18.21.1.4",
+                pespolicyid: "urn:oid:1.2.250.1.131.1.5.18.21.1.7",
+
                 pespolicydesc: "Politique de signature Helios de la DGFiP",
-                pespolicyhash: "Jkdb+aba0Hz6+ZPKmKNhPByzQ+Q=",
-                pesspuri: "https://portail.dgfip.finances.gouv.fr/documents/PS_Helios_DGFiP.pdf",
+
+                //pespolicyhash: "Jkdb+aba0Hz6+ZPKmKNhPByzQ+Q=",
+                pespolicyhash: "roF9+cfRHNPtVJolhdqfIqGMVuUXX8aR4rpiquf0u5E",
+
+                //pesspuri: "https://portail.dgfip.finances.gouv.fr/documents/PS_Helios_DGFiP.pdf",
+                pesspuri: "https://www.collectivites-locales.gouv.fr/files/files/finances_locales/dematerialisation/ps_helios_dgfip.pdf",
+
                 pescity: "<?php hecho($libersign_properties->get('libersign_city'))?>",
                 pespostalcode: "<?php hecho($libersign_properties->get('libersign_cp'))?>",
                 pescountryname: "France",
                 pesclaimedrole: "Ordonnateur",
                 pesencoding: "iso-8859-1",
-                format: "XADES-env"
+                //format: "XADES-env"
+                format: "xades-env-1.2.2-sha256"
             });
 
             $(".libersign").libersign({